Organize sort module
