+++ app/grid/[sortType]/[sortOrder]/page.tsx
@@ -6,8 +6,8 @@ import { cache } from 'react';
 import PhotoGridPage from '@/photo/PhotoGridPage';
 import { getDataForCategoriesCached } from '@/category/cache';
 import { getPhotosMetaCached } from '@/photo/cache';
-import { SortProps } from '@/photo/db/sort';
-import { getSortOptionsFromParams } from '@/photo/db/sort-path';
+import { SortProps } from '@/photo/sort';
+import { getSortOptionsFromParams } from '@/photo/sort/path';
 import { FEED_META_QUERY_OPTIONS, getFeedQueryOptions } from '@/feed';
 import { PhotoQueryOptions } from '@/photo/db';
