+++ src/photo/PhotoGridPageClient.tsx
@@ -10,7 +10,7 @@ import clsx from 'clsx/lite';
 import useElementHeight from '@/utility/useElementHeight';
 import MaskedScroll from '@/components/MaskedScroll';
 import { IS_RECENTS_FIRST } from '@/app/config';
-import { SortBy } from './db/sort';
+import { SortBy } from './sort';
 
 export default function PhotoGridPageClient({
   photos,
