+++ src/photo/PhotosLargeInfinite.tsx
@@ -3,7 +3,7 @@
 import { PATH_FULL_INFERRED } from '@/app/paths';
 import InfinitePhotoScroll from './InfinitePhotoScroll';
 import PhotosLarge from './PhotosLarge';
-import { SortBy } from './db/sort';
+import { SortBy } from './sort';
 
 export default function PhotosLargeInfinite({
   initialOffset,
